refactor: moved make-httprequest out of middleware into shared
